Generate Trelliscope Bundle

Generates a bundle of trelliscopes for publishing to the web

generate_trelliscope_bundle(include_image_panel = TRUE,
  include_timeline_info = TRUE, include_timeline_tone = FALSE,
  include_sentiment_bin = TRUE,
  wordcloud_modes = c("WordCloudImageTags", "WordCloudImageWebTags"),
  terms = NA, domains = NA, gkg_themes = NA, images_ocr = NA,
  images_tag = NA, images_web_tag = NA, images_face_tone = NA,
  images_num_faces = NA, images_web_count = NA,
  timespans = c("24 hours"), wordcloud_timespans = c("1 month"),
  dates = NA, source_countries = NA, tone = NA,
  tone_absolute_value = NA, maximum_records = 250, translate = NULL,
  timeline_smooth = 5, sort_by = "DateDesc", base_path = NULL,
  rows = 1, columns = 2, id_columns = NULL, group_columns = NULL,
  return_message = TRUE)

Arguments

include_image_panel
if TRUE includes ArtList

include_timeline_info
if TRUE includes TimelineInfo

include_timeline_tone
if TRUE includes Timeline tone

include_sentiment_bin
if TRUE includes sentiment bin

wordcloud_modes
Wordcloud items

NA - none

WordCloudNative - word cloud of native text for specified specified terms/domains/webtags/imagewebtags and OCR'd text

WordCloudImageTags - word cloud of resolved imagetags for specified terms/domains/webtags/imagewebtags and OCR'd text

WordCloudImageWebTags - word cloud of resolved image web tags for specified terms/domains/webtags/imagewebtags and OCR'd text

terms
a vector of terms

domains
a vector of webdomains

gkg_themes
global knowledge graph theme #' use get_gdelt_codebook_ft_api(code_book = "gkg")) for options

images_ocr
vector of words to search for OCR'd text

images_tag
vector of image tags from the image tag code book. use get_gdelt_codebook_ft_api(code_book = "imagetag")) for options

images_web_tag
vector of image tags from the image web tag code book. use get_gdelt_codebook_ft_api(code_book = "imageweb"))

images_face_tone
vector of facial tone scores

images_num_faces
vector of face count

images_web_count
numeric vector of number of times photo appeared

timespans
character vector of the time frame - no more than 12 weeks - default is 24 hours acceptable periods include:

hours (default)

minutes

weeks

months

wordcloud_timespans
timespans for wordcloud

dates
vector of dates in YMD HMS format, separated by - you can use the generate_dates() function to generate a vector of default NULL

source_countries
character source countries #' see get_gdelt_codebook_ft_api(code_book = "countries") for options

tone
numeric tone - default (NA)

tone_absolute_value
numeric tone absolute value (default NA)

maximum_records
Number between 1 and 250

timeline_smooth
if mode is a time-line

sort_by
sorting method

DateDesc - descending by date (default)

DateAsc - ascending by date

ToneDesc - descending tone

base_path
base path to save the trelliscopes

rows
rows of

return_message